+++ BioImageOperation/PathLink.cpp
@@ -49,11 +49,11 @@ double PathLink::getDirectionRate() {
 	return (double)(nNormal - nReverse) / getMax();
 }
 
-void PathLink::draw(Mat* image, Scalar color, int max, bool animate, int scale) {
-	int x1 = (int)(this->x1 * scale);
-	int y1 = (int)(this->y1 * scale);
-	int x2 = (int)(this->x2 * scale);
-	int y2 = (int)(this->y2 * scale);
+void PathLink::draw(Mat* image, Scalar color, int max, bool animate, double scaling, double scale) {
+	int x1 = (int)(this->x1 * scaling);
+	int y1 = (int)(this->y1 * scaling);
+	int x2 = (int)(this->x2 * scaling);
+	int y2 = (int)(this->y2 * scaling);
 	Point point;
 	double animPos2;
 
@@ -73,9 +73,9 @@ void PathLink::draw(Mat* image, Scalar color, int max, bool animate, int scale) 
 		point.x = (int)(x1 + animPos2 * (x2 - x1));
 		point.y = (int)(y1 + animPos2 * (y2 - y1));
 
-		drawMarker(*image, point, color, MarkerTypes::MARKER_CROSS, 2, 1, LineTypes::LINE_AA);
+		drawMarker(*image, point, color, MarkerTypes::MARKER_CROSS, (int)(2 * scale), (int)scale, LineTypes::LINE_AA);
 	} else {
-		line(*image, Point(x1, y1), Point(x2, y2), color, 1, LineTypes::LINE_AA);
+		line(*image, Point(x1, y1), Point(x2, y2), color, (int)scale, LineTypes::LINE_AA);
 	}
 }