+++ BioImageOperation/Util.cpp
@@ -655,7 +655,7 @@ string Util::combinePath(string basepath, string templatepath) {
     return (filesystem::path(basepath) / filesystem::path(templatepath)).string();
 }
 
-Size Util::drawText(Mat* image, string text, Point point, HersheyFonts fontFace, double fontScale0, Scalar color) {
+Size Util::drawText(Mat* image, string text, Point point, HersheyFonts fontFace, Scalar color, double fontScale0) {
 	double fontScale = fontScale0 * sqrt(image->cols / 1000.0 * image->rows / 1000.0) / 2;
 	int thickness = int(ceil(fontScale));
 	Size size = getTextSize(text, fontFace, fontScale, thickness, nullptr);
@@ -663,7 +663,7 @@ Size Util::drawText(Mat* image, string text, Point point, HersheyFonts fontFace,
 	return size;
 }
 
-void Util::drawAngle(Mat* image, double x, double y, double rad, double angle, Scalar color, bool isArrow) {
+void Util::drawAngle(Mat* image, double x, double y, double rad, double angle, Scalar color, bool isArrow, double scale) {
 	double radAngle = Util::degreesToRadians(angle);
 	double frontLength = 1;
 	if (isArrow) {
@@ -675,9 +675,9 @@ void Util::drawAngle(Mat* image, double x, double y, double rad, double angle, S
 	int y1 = (int)(y + frontLength * rad * sin(radAngle));
 
 	if (isArrow) {
-		arrowedLine(*image, Point(x0, y0), Point(x1, y1), color, 1, LineTypes::LINE_AA);
+		arrowedLine(*image, Point(x0, y0), Point(x1, y1), color, (int)scale, LineTypes::LINE_AA);
 	} else {
-		line(*image, Point(x0, y0), Point(x1, y1), color, 1, LineTypes::LINE_AA);
+		line(*image, Point(x0, y0), Point(x1, y1), color, (int)scale, LineTypes::LINE_AA);
 	}
 }